requirements.txt
torch
torchvision
torchaudio
torchtext
tensorboard
pyyaml
tqdm
numpy